fix(create-about): fix create about type
