+++ src/types/projectData.ts
@@ -6,21 +6,23 @@ export interface ProjectData {
   category: string;
   type: string;
   beneficiaries: string;
-  about: { // Change this to match the ProjectAboutData type
-    projectTitle: string,
-    projectDescription: string,
-    projectLink: string,
-    projectImage: string,
-    projectObjDescription: string,
-    projectObjImage: string,
-    projectName1: string,
-    projectName1Description: string,
-    projectName1Image: string,
-    projectName2: string,
-    projectName2Description: string,
-    projectName2Image: string,
-    theme: string,
-  },
+  about: ProjectAboutData;
   published: boolean;
   featured: boolean;
+}
+
+export interface ProjectAboutData {
+  projectTitle: string;
+  projectDescription: string;
+  projectLink: string;
+  projectImage: string;
+  projectObjDescription: string;
+  projectObjImage: string;
+  projectName1: string;
+  projectName1Description: string;
+  projectName1Image: string;
+  projectName2: string;
+  projectName2Description: string;
+  projectName2Image: string;
+  theme: string;
 }
